update: 5.3

+++ build.sh
@@ -1,8 +1,12 @@
 #!/bin/bash
-if [ -d "build" ] ; then
-    rm -rf build
+if [[ -z "$1" && "$1" = "total" ]]; then
+    if [ -d "build" ] ; then
+        rm -rf build
+    fi
+    mkdir build
 fi
-mkdir build
 cd build
-cmake3 ../
+if [[ -z "$1" && "$1" = "total" ]]; then
+    cmake3 ../
+fi
 make -j8
